refactor(import): move OpenClaw log parsing out of the source reader
`openclaw-source.ts` had already crossed the 300-line layout limit and
the multi-agent listing grew it further. The pure per-line parsing
(event split, JSON parse, message projection) has no file I/O and now
lives in `openclaw-log-events.ts`; the source keeps the paths, the
listing and the SQLite handle.
