+++ src/import/openclaw/openclaw-source.ts
@@ -3,6 +3,13 @@ import { Database as DatabaseCtor } from "../../native/load-better-sqlite3.js";
 import { existsSync, readdirSync, readFileSync, statSync } from "node:fs";
 import { join } from "node:path";
 
+import {
+  isoToMs,
+  parseLogLine,
+  projectLogMessage,
+  splitLogLines,
+} from "./openclaw-log-events.js";
+
 /**
  * Read-only access to a `~/.openclaw` state directory. This is the **only**
  * file in the OpenClaw import feature that touches OpenClaw's physical
@@ -194,8 +201,8 @@ export class OpenclawSource {
     let cwd: string | null = null;
     let model: string | null = null;
     let startedAtMs = 0;
-    for (const line of splitLines(text)) {
-      const event = parseLine(line);
+    for (const line of splitLogLines(text)) {
+      const event = parseLogLine(line);
       if (!event) continue;
       if (event.type === "session") {
         id = typeof event.id === "string" ? event.id : null;
@@ -220,10 +227,10 @@ export class OpenclawSource {
       throw new OpenclawSourceError(`failed to read ${meta.file}: ${message}`);
     }
     const messages: OpenclawMessage[] = [];
-    for (const line of splitLines(text)) {
-      const event = parseLine(line);
+    for (const line of splitLogLines(text)) {
+      const event = parseLogLine(line);
       if (!event || event.type !== "message") continue;
-      const projected = projectMessage(event);
+      const projected = projectLogMessage(event);
       if (projected) messages.push(projected);
     }
     return messages;
@@ -280,100 +287,4 @@ export class OpenclawSource {
       this.db = null;
     }
   }
-}
-
-interface RawEvent {
-  type?: string;
-  id?: unknown;
-  cwd?: unknown;
-  timestamp?: unknown;
-  modelId?: unknown;
-  message?: unknown;
-}
-
-function splitLines(text: string): string[] {
-  return text.split(/\r?\n/);
-}
-
-/** Parse one JSONL line; malformed or blank lines yield null (skipped). */
-function parseLine(line: string): RawEvent | null {
-  const trimmed = line.trim();
-  if (trimmed.length === 0) return null;
-  try {
-    const parsed = JSON.parse(trimmed);
-    if (parsed && typeof parsed === "object") return parsed as RawEvent;
-    return null;
-  } catch {
-    return null;
-  }
-}
-
-/** Project a raw `message` event into a neutral `OpenclawMessage`. */
-function projectMessage(event: RawEvent): OpenclawMessage | null {
-  const msg = event.message;
-  if (!msg || typeof msg !== "object") return null;
-  const m = msg as Record<string, unknown>;
-  const role = m.role;
-  if (role !== "user" && role !== "assistant" && role !== "toolResult") {
-    return null;
-  }
-  const blocks = projectBlocks(m.content);
-  const atMs =
-    typeof m.timestamp === "number"
-      ? Math.round(m.timestamp)
-      : isoToMs(event.timestamp) ?? 0;
-  return {
-    role,
-    blocks,
-    toolCallId: typeof m.toolCallId === "string" ? m.toolCallId : null,
-    toolName: typeof m.toolName === "string" ? m.toolName : null,
-    isError: m.isError === true,
-    atMs,
-  };
-}
-
-function projectBlocks(content: unknown): OpenclawBlock[] {
-  if (!Array.isArray(content)) return [];
-  const blocks: OpenclawBlock[] = [];
-  for (const raw of content) {
-    if (!raw || typeof raw !== "object") continue;
-    const block = raw as Record<string, unknown>;
-    switch (block.type) {
-      case "text":
-        if (typeof block.text === "string") {
-          blocks.push({ type: "text", text: block.text });
-        }
-        break;
-      case "thinking":
-        if (typeof block.thinking === "string") {
-          blocks.push({ type: "thinking", thinking: block.thinking });
-        }
-        break;
-      case "toolCall": {
-        const name = block.name;
-        if (typeof name !== "string" || name.length === 0) break;
-        blocks.push({
-          type: "toolCall",
-          id: typeof block.id === "string" ? block.id : null,
-          name,
-          args:
-            block.arguments && typeof block.arguments === "object" &&
-            !Array.isArray(block.arguments)
-              ? (block.arguments as Record<string, unknown>)
-              : {},
-        });
-        break;
-      }
-      default:
-        break;
-    }
-  }
-  return blocks;
-}
-
-/** Parse an ISO-8601 timestamp into integer ms; null on failure. */
-function isoToMs(value: unknown): number | null {
-  if (typeof value !== "string") return null;
-  const ms = Date.parse(value);
-  return Number.isNaN(ms) ? null : ms;
 }
